requirements.txt
# Data Handling
pandas==2.3.3
numpy==2.4.2

# Machine Learning
scikit-learn==1.6.1
matplotlib==3.8.4
xgboost==3.2.0
torch==2.10.0
pytorch-tabnet==4.1.0

# Visualization & UI
streamlit==1.54.0
altair==6.0.0

# Helpers
joblib==1.5.3
shap==0.45.1
fpdf2==2.7.8